Don't export promise timeout
