--- src/utils/promise.ts
@@ -1,40 +1,0 @@
-import fs from "fs";
-
-export const sleep = (ms: number): Promise<void> => new Promise((resolve): NodeJS.Timeout => setTimeout(resolve, ms))
-
-export const promiseTimeout = <T>(ms: number, promise: Promise<T>): Promise<T> => {
-  // Create a promise that rejects in <ms> milliseconds
-  const timeout = new Promise((resolve, reject): void => {
-    const id = setTimeout(() => {
-      clearTimeout(id);
-      reject(new Error(`Timed out after ${ms}ms.`))
-    }, ms)
-  })
-
-  // Returns a race between our timeout and the passed in promise
-  return Promise.race([promise, timeout]) as Promise<T>;
-}
-
-export const readDir = async (folderPath: string, options?: any): Promise<string[]> => {
-  return new Promise((resolve, reject): void => {
-    fs.readdir(folderPath, options, (err, files) => {
-      if (err) {
-        reject(err);
-      } else {
-        resolve(files);
-      }
-    });
-  })
-}
-
-export const writeFile = async (filePath: string, data: any): Promise<void> => {
-  return new Promise((resolve, reject): void => {
-    fs.writeFile(filePath, data, (err) => {
-      if (err) {
-        reject(err);
-      } else {
-        resolve();
-      }
-    });
-  });
-};
